refactor: Attendance에 name, timestamps 인스턴스 변수 추가
