+++ src/test/java/attendance/domain/AttendanceTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import java.time.LocalDateTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -100,5 +101,25 @@ public class AttendanceTest {
         assertThatCode(() -> Attendance.method2(24)).doesNotThrowAnyException();
     }
 
+    @DisplayName("출석 정보 저장 성공 ")
+    @Test
+    void test14() {
+        LocalDateTime localDateTime = LocalDateTime.of(2024, 12, 23, 13, 3);
+        Attendance attendance = new Attendance("빙봉");
 
+        assertThatCode(() -> attendance.add(localDateTime)).doesNotThrowAnyException();
+    }
+
+    @DisplayName("출석 정보 저장 실패")
+    @Test
+    void test15() {
+        LocalDateTime localDateTime = LocalDateTime.of(2024, 12, 23, 13, 3);
+        Attendance attendance = new Attendance("빙봉");
+
+        attendance.add(localDateTime);
+
+        assertThatThrownBy(() -> attendance.add(localDateTime))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessage("[ERROR] 출석 기록이 존재합니다. 출석 수정 기능을 이용하세요.");
+    }
 }
